Fix activity lifecycle bug
